selectImage

+++ src/modules/daybook/views/EntryView.vue
@@ -10,6 +10,14 @@ import { defineAsyncComponent } from 'vue';
         </div>
 
         <div>
+
+            <input type="file"
+                 @change="onSelectedImage"
+                 ref="imageSelector"
+                 v-show="false"
+                 accept="image/png, image/jpeg"
+            >
+
             <button 
             v-if="entry.id"
             class="btn btn-danger mx-2"
@@ -18,7 +26,8 @@ import { defineAsyncComponent } from 'vue';
                 <i class="fa fa-trash-all"></i>
             </button>
 
-            <button class="btn btn-primary">
+            <button class="btn btn-primary"
+                @click="onSelectImage">
                 Subir foto
                 <i class="fa fa-upload"></i>
             </button>
@@ -33,9 +42,15 @@ import { defineAsyncComponent } from 'vue';
         ></textarea>
     </div>
 
-    <img src="https://img.freepik.com/vector-gratis/fondo-plano-dia-internacional-gato_52683-86986.jpg" 
+    <!-- <img src="https://img.freepik.com/vector-gratis/fondo-plano-dia-internacional-gato_52683-86986.jpg" 
     alt="entry-picture"
-    class="img-thumbnail">
+    class="img-thumbnail"> -->
+
+    <img
+        v-if="localImage"
+        :src="localImage" 
+        alt="entry-picture"
+        class="img-thumbnail">
 
  </template>
 
@@ -51,6 +66,8 @@ import { defineAsyncComponent } from 'vue';
 <script>
 import {defineAsyncComponent} from 'vue'
 import { mapGetters, mapActions } from 'vuex';
+import Swal from 'sweetalert2'
+
 
 import getDayMonthYear from '../helpers/getDayMonthYear'
 
@@ -67,7 +84,9 @@ export default {
 
     data() {
         return {
-            entry: null
+            entry: null,
+            localImage: null,
+            file: null
         }
     },
     computed: {
@@ -109,6 +128,12 @@ export default {
         },
            async saveEntry(){
             
+            new Swal({
+                title: 'Espere por favor',
+                allowOutsideClick: false
+            })
+            Swal.showLoading()
+
             if(this.entry.id){
 
                 await this.updateEntry(this.entry)
@@ -118,16 +143,50 @@ export default {
 
                this.$router.push({name: 'entry', params: {id}})
             }
-
+            
+            Swal.fire('Guardado','Entrada registrada con éxito', 'success')
             
         },
         async onDeleteEntry() {
-            console.log('delete', this.entry)
+            const {isConfirmed} = await Swal.fire({
+                title: '¿Está seguro?',
+                text: 'Una vez borrado no se puede recuperar',
+                showDenyButton: true,
+                confirmButtonText: 'Si, estoy seguro'
+            })
 
-            await this.deleteEntry(this.entry.id)
+           if(isConfirmed){
+            new Swal({
+                title: 'Espere por favor',
+                allowOutsideClick: false
+            })
+            Swal.showLoading()
+
+            await this.deleteEntry(this.entry.id) 
 
             this.$router.push({name: 'no-entry'})
 
+            Swal.fire('Eliminado','','success')
+           }
+        },
+
+        onSelectedImage(event){
+           const file = event.target.files[0]
+
+           if (!file){
+            this.localImage = null
+            this.file = null
+                return
+           }
+
+            this.file = file
+
+           const fr = new FileReader()
+           fr.onload = () => this.localImage = fr.result
+           fr.readAsDataURL(file)
+        },
+        onSelectImage(){
+            this.$refs.imageSelector.click()
         }
     },
